Revert "fix: adjust agents creation and remove delay"

+++ application/src/main/java/agents/server/behaviour/AnnouncePowerRequest.java
@@ -91,7 +91,7 @@ public class AnnouncePowerRequest extends ContractNetInitiator {
 
     private ACLMessage chooseGreenSourceToExecuteJob(final List<ACLMessage> greenSourceOffers) {
         final Comparator<ACLMessage> compareGreenSources =
-                Comparator.comparingDouble(greenSource -> {
+                Comparator.comparingInt(greenSource -> {
                     try {
                         return getMapper().readValue(greenSource.getContent(), GreenSourceData.class).getAvailablePowerInTime();
                     } catch (final JsonProcessingException e) {
